Add setState to render properly, refactor a bit
