+++ lib/common.js
@@ -1,6 +1,6 @@
 import Airtable from 'airtable';
 
-import getEnvVars from '../environment_js';
+import getEnvVars from '../environment';
 
 // BASE is shared across our screens for Airtable API calls
 /**
