+++ screens/PointsHistoryScreen.js
@@ -1,58 +1,9 @@
-import React from 'react';
-import {
-  AsyncStorage,
-  ScrollView,
-  Text,
-  TouchableOpacity,
-  View
-} from 'react-native';
 import * as Font from 'expo-font';
-import BASE from '../lib/common';
-import Transactions from '../components/Transactions';
-Font.loadAsync({
-    Poppins: require('../assets/fonts/Poppins-Regular.ttf'),
-});
-const TransTable = BASE('Transactions').select({ view: 'Grid view' });
-// const CustTable = BASE('Customers').select({ view: 'Grid view' });
+import React from 'react';
+import { AsyncStorage, ScrollView, Text, View } from 'react-native';
 
-let transList = [];
-const getCustomerTransactions = function async(userId) {
-  return BASE('Transactions')
-    .select({ view: 'Transactions' })
-    .all()
-    .then(records => {
-      let userTrans = [];
-      //console.log(userId)
-      for (const record of records) {
-        const currCustomerId = record.get('Customer');
-        if (currCustomerId) {
-            //console.log(currCustomerId[0])
-            if (currCustomerId[0] === userId) {
-                userTrans.push(record);
-            }
-        }
-      }
-      //console.log(userTrans)
-      return Promise.all(userTrans);
-    })
-    .then(transactions => {
-      for (const transaction of transactions) {
-        const curr = {
-          date: new Date(transaction.get('Date')),
-          storeId: transaction.get('Store'),
-          productsPurchased: transaction.get('Products Purchased'),
-          points: transaction.get('Points Rewarded'),
-          id: transaction.get('transaction_id')
-        };
-        transList.push(curr);
-      }
-      transList.sort(function(a,b){
-        return new Date(b.date) - new Date(a.date);
-      })
-      console.log(transList)
-      return transList;
-    });
-};
+import Transactions from '../components/Transactions';
+import getCustomerTransactions from './historyHelpers';
 
 export default class PointsHistoryScreen extends React.Component {
   constructor(props) {
@@ -64,7 +15,14 @@ export default class PointsHistoryScreen extends React.Component {
 
   async componentDidMount() {
     const userId = await AsyncStorage.getItem('userId');
-    getCustomerTransactions(userId).then(console.log);
+    getCustomerTransactions(userId).then(transactions => {
+      // TODO @kyleqhua you'll want to split out the three most recent transations into a separate array; maybe call it 'recents'
+      this.setState({ transactions });
+    });
+
+    Font.loadAsync({
+      Poppins: require('../assets/fonts/Poppins-Regular.ttf')
+    });
   }
 
   render() {
@@ -72,18 +30,15 @@ export default class PointsHistoryScreen extends React.Component {
       <View>
         <Text> RECENT TRANSACTIONS</Text>
         <ScrollView>
+          {/* TODO @kyleqhua here, display recents first and add appropriate FE dividers etc */}
           {this.state.transactions.map(transaction => (
             <Transactions
               key={transaction.id}
               date={transaction.date}
               points={transaction.points}
-              storeId={transaction.storeId}
+              storeName={transaction.storeName}
             />
           ))}
-          <Transactions
-              date={new Date()}
-              points={111}
-              storeId={'Hi'}/>
         </ScrollView>
       </View>
     );
